# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VibeCare API
  version: 0.7.0
  description: Canonical mobile API. FITRUS credentials never cross this boundary.
servers:
  - url: https://api.example.invalid
    description: Replace per environment
security:
  - bearerAuth: []
paths:
...
  /v1/device-sessions:
    post:
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorizationId, deviceId]
              properties:
                authorizationId: { type: string, minLength: 1 }
                deviceId: { type: string, minLength: 1 }
      responses:
        '201': { description: Mock session started and ACK recorded }
        '200': { description: Same participant and identical request replayed with the same key }
        '400': { description: Missing or short Idempotency-Key, or invalid request body }
        '401': { description: Missing or invalid access token }
        '404': { description: Execution authorization not found for this participant and device }
        '409': { description: 'Expired or used authorization, conflicting key, busy target device, or changed algorithm' }
        '503': { description: Algorithm unavailable }
        '501': { description: Real device protocol is not configured }
  /v1/device-sessions/{id}/events:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType]
              properties:
                eventType:
                  type: string
                  enum: [ACK, RUNNING, STOPPING, COMPLETED, ERROR, DISCONNECTED]
                payload:
                  type: object
                  additionalProperties: true
                  default: {}
      responses:
        '202': { description: Simulator event accepted; COMPLETED closes the session }
        '400': { description: Invalid event type or payload }
        '401': { description: Missing or invalid access token }
        '404': { description: Session not found for the authenticated participant }
        '409': { description: Session already terminal }
  /v1/device-sessions/{id}/stop:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '200': { description: STOPPED or COMPLETED confirmed; terminal calls are idempotent }
        '400': { description: Missing or invalid stop reason }
        '401': { description: Missing or invalid access token }
        '404': { description: Session not found for the authenticated participant }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...